Master (#1)
* adds coc error gutter support

* remove duplicated entries

* Update colors/nord.vim

Co-Authored-By: Arctic Ice Studio <development@arcticicestudio.com>

* Fix typo

line 13

* Uniform status lines config for bundled airline and lightline themes (#169)

The included theme bundles have not supported the "uniform status line" feature (GH-58), which allows to change the background color of the status bar (StatusLine) group to `nord3`.

Related to GH-58
Resolves GH-168

* Plugin support for `vim-startify` (#176)

Added custom highlight groups of the `vim-startify` (1) plugin to adapt
to Nord's style.

References:
  (1) https://github.com/mhinz/vim-startify

Resolves GH-159

* Remove underline from gutter line numbers (#185)

Vim version 8.1.2029 [1] added the `underline` attribute for the
`CursorLineNr` group to `cterm` [2] based on vim/vim#4933 [3].
This change resulted in gutter line numbers being underlined which has
now been reverted back to Nord's style by explicitly setting the
attribute for the group to `NONE`.

[1]: https://github.com/vim/vim/releases/tag/v8.1.2029
[2]: https://github.com/vim/vim/compare/d9b0d83b13d2691e4544709abd87eac004715175...017ba07fa2cdc578245618717229444fd50c470d#diff-80fffb3e9c20e93e5b2328a9a20e19c9
[3]: https://github.com/vim/vim/pull/4933

Resolves GH-174

* Prepare release version 0.13.0

* adds coc error gutter support

* Add nvim-lsp support (#198)

Added highlighting support for the build-in Neovim language server [1] using the coc.nvim [2] groups as reference.

[1]: https://github.com/neovim/nvim-lsp
[2]: https://github.com/neoclide/coc.nvim

* Consistent `Error` and MoreMsg highlight group consistent between console and GUI modes. (#202)

Consistent `Error` and `MoreMsg` highligh. in term and GUI mode (#202)

Before the `Error` group in GUI mode used `nord0` as foreground color
instead of `nord4` resulting in a bad contrast.

Also after checking  ( links to it)
Also since there was also no color defined for terminal mode for the
`MoreMsg` group (see `:help MoreMsg` that link to `:help more-prompt`)
Vim used the default color which was some kind of green.
To ensure it matches Nord's style it has now been changed to use `nord8`
(main accent color) for both terminal and GUI mode.
This can be tested by running `:echon "MESSAGE\n"` taht produces a lot
of lines that won't fit on the current screen space anymore.

Co-authored-by: Arctic Ice Studio <development@arcticicestudio.com>
Co-authored-by: Sven Greb <development@svengreb.de>

* Use transparent background for gutter line number in GUI mode (#204)

The `LineNr` and `CursorLineNr` highlight groups now have a transparent
background in GUI mode.

Before it was set to `nord0_gui` which worked fine in most cases.
However, some plugins use these highlight groups to render their content
in a popup window which can potentially have a different background
color. This caused some issues e.g. for the fuzzy search plugin
LeaderF [1].

The compatibility with the `g:nord_cursor_line_number_background`
theme configuration has been verified to work as expected in both modes
when it is set to `0` or `1`.

This change is not related to the terminal mode or when using
`set notermguicolors` since `ctermbg` for `LineNr` and `CursorLineNr`
is set to `NONE` by default.

[1]: https://github.com/Yggdroot/LeaderF

* Release version 0.14.0

* adds coc error gutter support

* Add nvim-lsp support (#198)

Added highlighting support for the build-in Neovim language server [1] using the coc.nvim [2] groups as reference.

[1]: https://github.com/neovim/nvim-lsp
[2]: https://github.com/neoclide/coc.nvim

* Add basic TypeScript and improve TSX support (#208)

Basic highlighting support for TypeScript & TSX

Added basic support to highlight TypeScript & TSX syntax more
consistently through the HerringtonDarkholme/yats.vim plugin [1].
This includes improvements to highlight...

1. ...TypeScript interface an class names using `nord7` as foreground,
   where interfaces also use the bold attribute, to match with
   structs/classes.
2. ...global methods like e.g. `setTimeout` with `nord8` using the
   italic attribute to mark it kind of static. 
3. ...regular expressions with `nord13` as foreground color instead of
   the normal color for quoted strings (`nord14`) to make it easier to
   differ between both. 
4. ...global objects like `Error`, `JSON` and `console` with `nord7`.
5. ...primitive/builtin types like `string` with `nord9`.
6. ...TypeScript type references with `nord7`.
7. ...TypeScript specific characters like for type annotations (`:`) and
   member optionality (`?`) as operator with `nord9`.

This also includes improvements for "vanilla" JavaScript elements.

[1]: https://github.com/HerringtonDarkholme/yats.vim

Resolves GH-208

Co-authored-by: Arctic Ice Studio <development@arcticicestudio.com>
Co-authored-by: Sven Greb <development@svengreb.de>

* Add support for vim-clap (#178)

Added basic support for vim-clap [1], a modern and performant generic finder and dispatcher for Vim and NeoVim.

[1]: https://github.com/liuchengxu/vim-clap

GH-178

Co-authored-by: Arctic Ice Studio <development@arcticicestudio.com>
Co-authored-by: Sven Greb <development@svengreb.de>

* Release version 0.15.0

Co-authored-by: Arctic Ice Studio <development@arcticicestudio.com>
Co-authored-by: john.hennessey <john.hennessey@ef.com>
Co-authored-by: John Hennessey <jghennes@gmail.com>
Co-authored-by: Radu Vasilescu <vasilescur@gmail.com>
Co-authored-by: Jose M. Murinello <jm.murinello@gmail.com>
Co-authored-by: Alexander Jeurissen <1220084+alexanderjeurissen@users.noreply.github.com>
Co-authored-by: xulongwu4 <xulongwu4@gmail.com>
Co-authored-by: Sven Greb <development@svengreb.de>
Co-authored-by: iamdi <helloiamdi@gmail.com>
Co-authored-by: Johan <meck@users.noreply.github.com>